requirements.txt
requests
markdown
beautifulsoup4
python-docx
python-dotenv
feedparser
msal
schedule
